requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-generativeai==0.8.5
Pillow==11.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
